added classes and game
made a small target game. Moved Face into its own class. Tracker was moved into its own class (the model, the window, and the cam). A bug was introduced, sometimes the target is hit when the end effector is not near it. Noticed that when reading from landmark file, this is not done from root folder (should be change in other files as well

related to issue #35